titan: Add falcon and peregrine sensors HAL
This sensors HAL supports both falcon and peregrine. The support for
the gyroscope is included depending on the target device.

It supports all the sensors supported by the stock sensors HAL with
the addition of a Significant Motion sensor. Since the Display Rotation
sensor and the Significant Motion sensor are mutually exclusive, the
HAL doesn't expose the Display Rotation sensor to avoid conflicts,
unless DISPLAY_ROTATION_SENSOR_ENABLED is set to true.

Kernel changes are required for this sensors HAL to work properly.

Change-Id: I1130bff103f3efdbb7315c068ea4b81d8f6b8955