titan: Restore QCOM time daemon
This restores the proprietary time daemon, which was removed as it was
believed it was the cause of date and time config losses some users
experienced. However, the Sony timekeeper is not as well integrated
with the framework as the QCOM one and actually causes issues such as
wrong device uptime.

This reverts commits 73fe478db58ea134ff6d66df751d48be5c0ca312 and
a156c26e06807e5f2d87c7b6db1f8f2164d897c5 without bringing back the
unneeded sepolicy/platform_app.te.

Change-Id: I385087b87cf9db285d00cbd1b26e626040c1ef49