titan: Bring back the common camera wrapper
Move titan and thea fixes in the common wrapper and add a header for
extra camaera parameters.

Also, prevent the vendor HAL from setting unsupported fps preview
ranges when changing certain camera parameters on falcon and peregrine
and disable the HW face detection since it simply causes a stream of
errors when enabled.

Change-Id: I5e0699ec74da7d53690a2706feddb2c614279421